@@ -133,20 +133,16 @@ var Spy = function (_React$Component) {
   }
 
   _createClass(Spy, [{
-    key: 'componentWillMount',
-    value: function componentWillMount() {
-      var history = this.props.history;
-
-      globalHistory = history;
-    }
-  }, {
-    key: 'componentWillReceiveProps',
-    value: function componentWillReceiveProps(nextProps) {
-      globalHistory = nextProps.history;
-    }
-  }, {
     key: 'render',
     value: function render() {
+      return null;
+    }
+  }], [{
+    key: 'getDerivedStateFromProps',
+    value: function getDerivedStateFromProps(nextProps) {
+      if (globalHistory !== nextProps.history) {
+        globalHistory = nextProps.history;
+      }
       return null;
     }
   }]);
